webapp/requirements.txt
numpy
python-crontab
flask
flask-SocketIO
pymysql
cryptography